Refactor Grid component by moving helper functions to utils file
